initial commit move template implementation code to .h file, rename internal constants variable
