+++ rlclientlib/rlclientlib.vcxproj
@@ -299,8 +299,7 @@ $(flatcPath) -o "$(SolutionDir)rlclientlib\generated\v2" --cpp "$(SolutionDir)rl
     <ClInclude Include="utility\http_helper.h" />
     <ClInclude Include="utility\eventhub_http_authorization.h" />
     <ClInclude Include="utility\apim_http_authorization.h" />
-    <ClCompile Include="azure_factories.cc" />
-    <ClCompile Include="logger\http_transport_client.cc" />
+    <ClCompile Include="azure_factories.cc" />    
     <ClCompile Include="model_mgmt\restapi_data_transport.cc" />
     <ClCompile Include="utility\http_client.cc" />
     <ClCompile Include="utility\http_helper.cc" />
